feat: Add 'edit profile' section on settings page
On this section of the page, users can edit their profile picture, name, username, and bio.

Editing profile picture:
    - Using Cloudinary API to host the images. Links to the images are stored in the database.
    - Using react-easy-crop package to allow users to crop image to a square shape and a set size.
    This will provide consistency to the shapes and sizes of the images.
    - All users start with a default profile picture.
    - If the user removes a custom profile picture, they will have the default.
    - File upload only allows images (jpeg or png)
    - If the user updates their profile picture and saves those changes, the previous image will be removed from the Cloudinary as well.

Editing name, username, and bio (form validation):
    - Each field has a max character limit.
    - Username cannot contain white spaces or special characters.
    - If username is changed, it still checks if new username is not taken.
    - Inserted text below the bio text field to show how many characters are remaining.
    - Bio can have a maximum of 250 characters.
    - Name can have a maximum of 30 characters.
    - Username can have a maximum of 30 characters.

Successful form submission:
    - A modal will appear once the changes have been successfully updated in the database, and the page will refresh upon closing the modal.
    - Refreshing the page after submitting changes helps ensure all values are up to date accross all pages.
    - User can cancel changes (before saving) to reset everything to their initial values.

Other Changes:
    - Updated username field on registration page to also not allow special characters or white spaces.
    - Created a secrets file.

Address: Issues #12, #1

+++ springboot-backend/src/main/java/com/example/springbootbackend/model/EmailVerificationToken.java
@@ -10,7 +10,7 @@ public class EmailVerificationToken implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Integer id;
 
     @OneToOne
     @JoinColumn(name = "email", referencedColumnName = "email", unique = true, nullable = false)
@@ -37,11 +37,11 @@ public class EmailVerificationToken implements Serializable {
         this.expiresAt = expiresAt;
     }
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
